clp-backend/requirements.txt
langgraph>=0.2
langchain>=0.2
langchain-community>=0.2
pydantic>=2.0
rapidfuzz>=3.0
typer>=0.12
typing-extensions>=4.8
google-generativeai>=0.7
google-genai>=0.3.0
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0

